+++ apps/website/src/app/i18n/messages/index.ts
@@ -1,12 +1,17 @@
 import type { WebLocale } from '../locale.types';
 import type { WebMessages } from '../web-messages.types';
+import { de } from './de';
 import { el } from './el';
 import { en } from './en';
 import { esES } from './es-ES';
 import { frFR } from './fr-FR';
+import { is } from './is';
+import { nl } from './nl';
 import { pl } from './pl';
+import { ptBR } from './pt-BR';
+import { zhCN } from './zh-CN';
 
-export { el, en, esES, frFR, pl };
+export { de, el, en, esES, frFR, is, nl, pl, ptBR, zhCN };
 
 export const WEB_MESSAGES: Record<WebLocale, WebMessages> = {
   en,
@@ -14,4 +19,9 @@ export const WEB_MESSAGES: Record<WebLocale, WebMessages> = {
   el,
   pl,
   'es-ES': esES,
+  de,
+  'pt-BR': ptBR,
+  'zh-CN': zhCN,
+  is,
+  nl,
 };